+++ src/main/dto/EventEntryCsv.ts
@@ -1,0 +1,16 @@
+export interface EventEntryCsv {
+  eventEntryId: string;
+  eventType: string;
+  start: string;
+  end: string;
+  summary: string;
+  projectId: string;
+  projectName: string;
+  categoryId: string;
+  categoryName: string;
+  taskId: string;
+  taskName: string;
+  labelIds: string;
+  labelNames: string;
+  description: string;
+}
